requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
Flask-JWT
Flask-Cors
kafka-python
pyspark
